requirements.txt
altair==4.2.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.3
debugpy==1.6.7
decorator==4.4.2
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
lit==16.0.5
llvmlite==0.40.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numba==0.57.0
numpy==1.24.3
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments==2.15.1
pymongo==4.3.3
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.22.0
streamlit-aggrid==0.3.4.post3
streamlit-scrollable-textbox==0.0.3
sympy==1.11.1
tenacity==8.2.2
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
torch
torchaudio
torchvision
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@d1c6c5c4d618fa950813c0c71aede34a5ac851e9
zipp==3.15.0
